feat(component): add Collapse component
