--- packages/big-design/src/components/Accordion/index.ts
@@ -1,4 +1,0 @@
-import { AccordionProps as _AccordionProps } from './Accordion';
-
-export { Accordion } from './Accordion';
-export type AccordionProps = _AccordionProps;
