+++ packages/big-design/src/components/Collapse/index.ts
@@ -1,0 +1,4 @@
+import { CollapseProps as _CollapseProps } from './Collapse';
+
+export { Collapse } from './Collapse';
+export type CollapseProps = _CollapseProps;
