+++ packages/docs/PropTables/index.ts
@@ -1,8 +1,8 @@
-export * from './AccordionPropTable';
 export * from './BadgePropTable';
 export * from './BoxPropTable';
 export * from './ButtonPropTable';
 export * from './CheckboxPropTable';
+export * from './CollapsePropTable';
 export * from './CounterPropTable';
 export * from './DisplayPropTable';
 export * from './DropdownPropTable';
