+++ packages/docs/next.config.js
@@ -20,13 +20,13 @@ module.exports = {
   },
   exportPathMap: () => ({
     '/': { page: '/GettingStarted/GettingStartedPage' },
-    '/accordion': { page: '/Accordion/AccordionPage' },
     '/alert': { page: '/Alert/AlertPage' },
     '/badge': { page: '/Badge/BadgePage' },
     '/box': { page: '/Box/BoxPage' },
     '/breakpoints': { page: '/Breakpoints/BreakpointsPage' },
     '/button': { page: '/Button/ButtonPage' },
     '/checkbox': { page: '/Checkbox/CheckboxPage' },
+    '/collapse': { page: '/Collapse/CollapsePage' },
     '/counter': { page: '/Counter/CounterPage' },
     '/colors': { page: '/Colors/ColorsPage' },
     '/display': { page: '/Display/DisplayPage' },
